Move mapif functions from int_storage.c to mapif.c
